requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
docker==6.1.3
exceptiongroup==1.2.2
filelock==3.18.0
Flask==2.3.2
frozenlist==1.5.0
fsspec==2025.3.2
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.314
langsmith==0.0.92
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.72.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
PyPDF2==3.0.1
pytest==7.4.0
python-dotenv==1.1.0
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.13.3
tomli==2.2.1
torch==2.0.1
tqdm==4.67.1
transformers==4.30.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.19.0
zipp==3.21.0
